feat(radar): score com 1 casa decimal para melhor filtragem de vagas
- Motor de score (scoring.js): Math.round → arredondamento com 1 decimal
  nas funções scoreVaga, computeReverseFit e computeAlignmentScore
- Migration 055: colunas fit_score*, reverse_fit_score, alignment_score
  migradas de INTEGER para NUMERIC(3,1)
- Parser IA (radar-prompt.js): prompt e parseAnalysisJson aceitam decimal
- MCP (radar-server.js): remove .int() do schema Zod de save_analysis
- UI admin (admin-core.js): radarBadge e pontos de exibição usam toFixed(1)
- CLI (jobsync.mjs): scoreColor e --min-score suportam valores decimais
- Testes: atualiza asserção de score inteiro e adiciona caso decimal

+++ mcp/radar-server.js
@@ -87,7 +87,7 @@ server.registerTool('save_analysis',
     { title: 'Salvar análise', description: 'Grava o resultado da análise em um lead (status → avaliada).',
       inputSchema: {
           id: z.string(),
-          fit_score: z.number().int().min(0).max(10).nullable().optional(),
+          fit_score: z.number().min(0).max(10).nullable().optional(),
           required_keywords: z.array(z.string()).optional(),
           nice_to_have_keywords: z.array(z.string()).optional(),
           gaps: z.array(z.string()).optional(),
